react refactoring
