--- draft-project/src/components/main/main.css
@@ -1,83 +1,0 @@
-/* web fonts */
-@import url('http://fonts.googleapis.com/earlyaccess/notosanskr.css');
-
-/* custom variables */
-:root {
-  --deep-orange: #D9493D;
-  --soft-orange: #F52423;
-}
-
-/* HTML tags */
-ul {
-  list-style: none;
-  padding-left: 0;
-}
-
-main > div {
-  display: inline;
-  float: left;
-}
-
-/* classes */
-.soft-orange {
-  color: var(--soft-orange);
-}
-
-.deep-orange {
-  color: var(--deep-orange);
-}
-
-.main-ui {
-  width: 80%;
-  margin: auto;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-
-.main-header {
-  text-align: center;
-  margin-top: 6em;
-  margin-bottom: 6em;
-}
-
-.left-info {
-  min-width: 20em;
-}
-
-.left-info > section {
-  width: 50%;
-  border-bottom: solid gray 1px;
-  padding-bottom: 1em;
-  margin-bottom: 1em;
-}
-
-.left-info > section > ul {
-  font-size: 1.3em;
-  line-height: 1.3em;
-}
-
-.right-articles .exercise {
-  margin-top: 3em;
-  margin-bottom: 4.5em;
-}
-.right-articles p {
-  margin-bottom: 3em;
-}
-
-/* links style */
-a:link {
-  text-decoration: none;
-  color: black;
-}
-
-a:visited {
-  color: black;
-}
-
-a:hover {
-  color: orange;
-}
-
-a:active {
-  text-decoration: none;
-}
-
