+++ orange/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "draft-project",
+  "name": "orange",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
