+++ orange/src/App.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Main } from './containers';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Main />
+    </div>
+  );
+}
+
+export default App;
