+++ orange/src/components/Article/Article.tsx
@@ -1,0 +1,6 @@
+// import React from 'react';
+
+export interface ArticleProps {
+    title?: string;
+    className?: string;
+}
