+++ orange/src/components/Articles/Articles.css
@@ -1,0 +1,8 @@
+.right-articles .exercise {
+  margin-top: 3em;
+  margin-bottom: 4.5em;
+}
+
+.right-articles p {
+  margin-bottom: 3em;
+}
