+++ orange/src/components/Header/Header.css
@@ -1,0 +1,6 @@
+.main-header {
+  text-align: center;
+  margin-top: 6em;
+  margin-bottom: 6em;
+}
+
