+++ orange/src/components/index.tsx
@@ -1,0 +1,9 @@
+import Header from './Header/Header';
+import StudyInfo from './StudyInfo/StudyInfo';
+import Articles from './Articles/Articles';
+
+export {
+    Header,
+    StudyInfo,
+    Articles
+};
