+++ orange/src/containers/index.tsx
@@ -1,0 +1,5 @@
+import Main from './Main/Main';
+
+export {
+    Main
+};
